Added support for using RSpec's argument matchers with Not A Mock's ArgsMatcher.
Extended RSpec's argument matchers with an in_any_order matcher, to match array arguments without caring about order of elements in the array. This is particularly useful for matching against arrays of database records, where you're not sure of the order in which the records have come out of the database.